requirements-dev.txt
pytest >= 5.4
pytest-click
coverage >= 5.2
pytest-datafiles
